[EasyCodingStandard] migrate RunCommand responsibility to Configuration
